# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tob_consent_status/schema": {
      "get": {
        "description": "Retrive the datatype of the tob_consent_status table",
        "operationId": "tob_consent_statusController_findSchema",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consent_statusEntity"
                }
              }
            }
          }
        },
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consent_status/get": {
      "get": {
        "description": "Filter all the records from the tob_consent_status table",
        "operationId": "tob_consent_statusController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_consent_statusEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consent_status/{id}": {
      "get": {
        "description": "Read all the records from the tob_consent_status table",
        "operationId": "tob_consent_statusController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consent_statusEntity"
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the tob_consent_status table",
        "operationId": "tob_consent_statusController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updatetob_consent_statusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consent_statusEntity"
                }
              }
            }
          }
        },
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the tob_consent_status table",
        "operationId": "tob_consent_statusController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consent_statusEntity"
                }
              }
            }
          }
        },
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consent_status/query": {
      "post": {
        "description": "Read all the records from the tob_consent_status table",
        "operationId": "tob_consent_statusController_findAllwithquery",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Querytob_consent_statusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_consent_statusEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consent_status": {
      "get": {
        "description": "Read all the records from the tob_consent_status table",
        "operationId": "tob_consent_statusController_findAll",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_consent_statusEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the tob_consent_status table",
        "operationId": "tob_consent_statusController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Createtob_consent_statusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consent_statusEntity"
                }
              }
            }
          }
        },
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consents/schema": {
      "get": {
        "description": "Retrive the datatype of the tob_consents table",
        "operationId": "tob_consentsController_findSchema",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consentsEntity"
                }
              }
            }
          }
        },
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consents/get": {
      "get": {
        "description": "Filter all the records from the tob_consents table",
        "operationId": "tob_consentsController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_consentsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consents/{id}": {
      "get": {
        "description": "Read all the records from the tob_consents table",
        "operationId": "tob_consentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consentsEntity"
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the tob_consents table",
        "operationId": "tob_consentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updatetob_consentsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consentsEntity"
                }
              }
            }
          }
        },
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the tob_consents table",
        "operationId": "tob_consentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_consentsEntity"
                }
              }
            }
          }
        },
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_consents/query": {
      "post": {
        "description": "Read all the records from the tob_consents table",
        "operationId": "tob_consentsController_findAllwithquery",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Querytob_consentsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_consentsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tgadev.gsstvl.com/ct242/tob001/tob002/v1/api-int",
      "description": "Production Server"
    }
  ],
  "components": {
    "schemas": {
...
      "tob_consent_lfiEntity": {
        "type": "object",
        "properties": {
          "lfi_code": {
            "type": "string"
          },
          "lfi_name": {
            "type": "string"
          },
          "onbehalfof": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tob_consent_onbehalfofEntity"
            }
          }
        },
        "required": [
          "lfi_code",
          "lfi_name",
          "onbehalfof"
        ]
      },
      "tob_consent_statusEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "lfi": {
            "$ref": "#/components/schemas/tob_consent_lfiEntity"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "lfi",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
      "Querytob_consent_lfiDto": {
        "type": "object",
        "properties": {
          "lfi_code": {
            "type": "string"
          },
          "lfi_name": {
            "type": "string"
          },
          "onbehalfof": {
            "$ref": "#/components/schemas/Querytob_consent_onbehalfofWapperDto"
          }
        },
        "required": [
          "lfi_code",
          "lfi_name",
          "onbehalfof"
        ]
      },
      "Querytob_consent_lfiWapperDto": {
        "type": "object",
        "properties": {
          "is": {
            "$ref": "#/components/schemas/Querytob_consent_lfiDto"
          }
        },
        "required": [
          "is"
        ]
      },
      "Querytob_consent_statusDto": {
        "type": "object",
        "properties": {
          "lfi": {
            "$ref": "#/components/schemas/Querytob_consent_lfiWapperDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "lfi",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
      "Createtob_consent_statusDto": {
        "type": "object",
        "properties": {
          "lfi": {
            "$ref": "#/components/schemas/Createtob_consent_lfiDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "lfi",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
      "Updatetob_consent_lfiDto": {
        "type": "object",
        "properties": {
          "lfi_code": {
            "type": "string"
          },
          "lfi_name": {
            "type": "string"
          },
          "onbehalfof": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Updatetob_consent_onbehalfofDto"
            }
          }
        },
        "required": [
          "lfi_code",
          "lfi_name",
          "onbehalfof"
        ]
      },
      "Updatetob_consent_statusDto": {
        "type": "object",
        "properties": {
          "lfi": {
            "$ref": "#/components/schemas/Updatetob_consent_lfiDto"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "lfi",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "tob_consentsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "requesturl": {
            "type": "string"
          },
          "requesttimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "type": "string"
          },
          "baseconsentid": {
            "type": "string"
          },
          "resourceconsentid": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "isauthorized": {
            "type": "boolean"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "requesturl",
          "requesttimestamp",
          "method",
          "baseconsentid",
          "resourceconsentid",
          "consentgroupid",
          "consenttype",
          "requestdata",
          "responsedata",
          "isauthorized",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
      "Updatetob_consentsDto": {
        "type": "object",
        "properties": {
          "requesturl": {
            "type": "string"
          },
          "requesttimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "type": "string"
          },
          "baseconsentid": {
            "type": "string"
          },
          "resourceconsentid": {
            "type": "string"
          },
          "consentgroupid": {
            "type": "string"
          },
          "consenttype": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "isauthorized": {
            "type": "boolean"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "requesturl",
          "requesttimestamp",
          "method",
          "baseconsentid",
          "resourceconsentid",
          "consentgroupid",
          "consenttype",
          "requestdata",
          "responsedata",
          "isauthorized",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
...
    }
  }
}